# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateNotification:
    properties:
      message:
        example: You have a new message
        type: string
      status:
        example: new
        type: string
      title:
        example: New Message
        type: string
    type: object
  model.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  model.Notification:
    properties:
      id:
        example: 1
        type: integer
      message:
        example: You have a new message
        type: string
      status:
        example: new
        type: string
      title:
        example: New Message
        type: string
    type: object
  model.Order:
    properties:
      created_at:
        type: string
      customer_name:
        type: string
      id:
        type: integer
      order_products:
        items:
          $ref: '#/definitions/model.OrderProduct'
        type: array
      payment_method:
        type: integer
      status:
        type: string
      table_id:
        type: integer
      tax:
        type: number
      total_amount:
        type: number
      updated_at:
        type: string
    required:
    - customer_name
    - table_id
    type: object
  model.OrderProduct:
    properties:
      id:
        type: integer
      order_id:
        type: integer
      product_id:
        type: integer
      qty:
        type: integer
    type: object
  model.OrderProductResponse:
    properties:
      item:
        type: string
      order_id:
        type: integer
      price:
        type: number
      qty:
        type: integer
    type: object
  model.OrderResponse:
    properties:
      customer_name:
        type: string
      id:
        type: integer
      order_date:
        type: string
      order_products:
        items:
          $ref: '#/definitions/model.OrderProductResponse'
        type: array
      status:
        type: string
      sub_total:
        type: integer
      table_id:
        type: integer
    type: object
  model.Payment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://example.com/support
  description: API for Point Of Sale
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: App pos Team 2
  version: "1.0"
paths:
...
  /notification:
    get:
      consumes:
      - application/json
      description: Retrieve all notifications, optionally filtered by status
      parameters:
      - description: Notification status (e.g., unread)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of notifications retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Notification'
                  type: array
              type: object
        "500":
          description: Failed to fetch notifications
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Get all notifications
      tags:
      - Notification
    post:
      consumes:
      - application/json
      description: Create a new notification
      parameters:
      - description: Notification payload
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/model.CreateNotification'
      produces:
      - application/json
      responses:
        "201":
          description: Notification created successfully
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Notification'
              type: object
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to create notification
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Create a new notification
      tags:
      - Notification
  /notification/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a notification using its ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Notification deleted successfully
          schema:
            $ref: '#/definitions/model.SuccessResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to delete notification
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Delete a notification by ID
      tags:
      - Notification
    get:
      consumes:
      - application/json
      description: Retrieve a notification by its ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Notification retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Notification'
              type: object
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to fetch notification
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Get a notification by ID
      tags:
      - Notification
    put:
      consumes:
      - application/json
      description: Update the status or details of a notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Notification updated successfully
          schema:
            $ref: '#/definitions/model.SuccessResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to update notification
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Update a notification by ID
      tags:
      - Notification
  /notification/mark-all-read:
    put:
      consumes:
      - application/json
      description: Mark all notifications as read
      produces:
      - application/json
      responses:
        "200":
          description: All notifications marked as read successfully
          schema:
            $ref: '#/definitions/model.SuccessResponse'
        "500":
          description: Failed to mark notifications as read
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Mark all notifications as read
      tags:
      - Notification
  /order:
    get:
      consumes:
      - application/json
      description: Retrieve a list of orders with optional search and status filtering.
      parameters:
      - description: Search keyword to filter orders
        in: query
        name: search
        type: string
      - description: Filter orders by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order successfully retrieved
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.OrderResponse'
                  type: array
              type: object
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Retrieve all orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order with associated products
      parameters:
      - description: Order payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Order successfully created
          schema:
            $ref: '#/definitions/model.SuccessResponse'
        "400":
          description: Failed to create order
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Invalid input
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Create a new order
      tags:
      - Orders
  /order/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted order
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Delete an order
      tags:
      - Orders
    put:
      consumes:
      - application/json
      description: Update the details of an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated order payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.Order'
      produces:
      - application/json
      responses:
        "200":
          description: Order successfully updated
          schema:
            $ref: '#/definitions/model.SuccessResponse'
        "400":
          description: Failed to update order
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Invalid input
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Update an existing order
      tags:
      - Orders
  /order/payment:
    get:
      consumes:
      - application/json
      description: Get a list of all payment methods
      produces:
      - application/json
      responses:
        "200":
          description: Payment methods successfully retrieved
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Payment'
                  type: array
              type: object
        "404":
          description: Payment methods not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Authentication: []
      summary: Retrieve all payment methods
      tags:
      - Payments
...
securityDefinitions:
  Authentication:
    in: header
    name: Authorization
    type: apiKey
  UserID:
    in: header
    name: User-ID
    type: apiKey
  UserRole:
    in: header
    name: User-Role
    type: apiKey
swagger: "2.0"